Several significant changes.
Changed the location of the header "H1" of the site header on the desktop version.
Added a js script to replace the content of h1 and h2 of the product category page with data from the database.
Removed (commented out) sections "Blog", "About Us" and "Frequently Asked Questions" from the main menu.
Replaced the photo of the header of the product category page (added blackout for each one so that the header text is read better).