Added the ability to send orders.
Added a script for sending data from the order form to the server and to the telegram bot. In the script, I changed the way of generating a link to a product.
Added a new path to url.conf path('makeorder/', makeorder, name='makeorder').
Added application appOrders.
Added appTelegram app.
Edited the 404 error page template.