Set up the admin panel. Added goods.
Added filters, search and a panel to the top with save buttons in the admin panel for the appProductItem application

Added one item each in the category "Pen cases" and "Bags".
Updated category pictures for accessories, bags, pen cases, money clips, and others.

I figured out the settings for saving static and media files.