Fixed several bugs.
Made the following changes to the meta tags: changed the name of the workshop to "Qualitas leather", added an image for the main page, contacts, delivery methods, payment methods, fixed the display of the image for the product page.
Added the CSRF_TRUSTED_ORIGINS instruction to the settings.py file for the admin panel to work correctly on the prod server.